requirements.txt
asgiref==3.3.4
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.12.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-renderer-xlsx==0.4.0
et-xmlfile==1.1.0
future==0.18.2
idna==2.10
mypy-extensions==0.4.3
netmiko==3.4.0
nornir==3.1.1
nornir-netmiko==0.1.1
nornir-utils==0.1.2
ntc-templates==2.0.0
openpyxl==3.0.7
paramiko==2.7.2
pycparser==2.20
PyJWT==1.7.1
PyMySQL==1.0.2
PyNaCl==1.4.0
pyserial==3.5
python-gitlab==2.8.0
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
scp==0.13.4
six==1.16.0
sqlparse==0.4.1
tenacity==7.0.0
textfsm==1.1.0
typing-extensions==3.10.0.0
urllib3==1.26.5
